+++ src/drivers/cmakeDriver.ts
@@ -16,7 +16,7 @@ import { CMakeOutputConsumer } from '@cmt/diagnostics/cmake';
 import { RawDiagnosticParser } from '@cmt/diagnostics/util';
 import { ProgressMessage } from '@cmt/drivers/drivers';
 import * as expand from '@cmt/expand';
-import { CMakeGenerator, effectiveKitEnvironment, Kit, kitChangeNeedsClean, KitDetect, getKitDetect, getVSKitEnvironment } from '@cmt/kits/kit';
+import { CMakeGenerator, effectiveKitEnvironment, Kit, kitChangeNeedsClean, KitDetect, getKitDetect, getVSKitEnvironment, getVsKitPreferredGenerator } from '@cmt/kits/kit';
 import * as logging from '@cmt/logging';
 import paths from '@cmt/paths';
 import { fs } from '@cmt/pr';
@@ -376,7 +376,11 @@ export abstract class CMakeDriver implements vscode.Disposable {
 
             return envs;
         } else {
-            return {};
+            let envs = this._kitEnvironmentVariables;
+            envs = EnvironmentUtils.merge([envs, await this.computeExpandedEnvironment(this.config.environment, envs)]);
+            envs = EnvironmentUtils.merge([envs, await this.computeExpandedEnvironment(this.config.cpackEnvironment, envs)]);
+            envs = EnvironmentUtils.merge([envs, await this.computeExpandedEnvironment(this._variantEnv, envs)]);
+            return envs;
         }
     }
 
@@ -752,8 +756,14 @@ export abstract class CMakeDriver implements vscode.Disposable {
         this._kitEnvironmentVariables = await effectiveKitEnvironment(kit, this.expansionOptions);
 
         // Place a kit preferred generator at the front of the list
-        if (kit.preferredGenerator) {
-            preferredGenerators.unshift(kit.preferredGenerator);
+        // For VS kits that don't have preferredGenerator (e.g., scanned before a VS version was added),
+        // try to derive it from the VS installation.
+        let kitPreferredGenerator = kit.preferredGenerator;
+        if (!kitPreferredGenerator && kit.visualStudio) {
+            kitPreferredGenerator = await getVsKitPreferredGenerator(kit);
+        }
+        if (kitPreferredGenerator) {
+            preferredGenerators.unshift(kitPreferredGenerator);
         }
 
         // If no preferred generator is defined by the current kit or the user settings,
@@ -838,14 +848,14 @@ export abstract class CMakeDriver implements vscode.Disposable {
             const opts = this.expansionOptions;
             opts.envOverride = await this.getConfigureEnvironment(configurePreset);
 
+            const installPrefix = this.config.installPrefix;
+            if (installPrefix) {
+                this._installDir = util.lightNormalizePath(await expand.expandString(installPrefix, opts));
+            }
+
             if (!this.useCMakePresets) {
                 const scope = this.workspaceFolder ? vscode.Uri.file(this.workspaceFolder) : undefined;
                 this._binaryDir = util.lightNormalizePath(await expand.expandString(this.config.buildDirectory(this.isMultiProject, scope), opts));
-
-                const installPrefix = this.config.installPrefix;
-                if (installPrefix) {
-                    this._installDir = util.lightNormalizePath(await expand.expandString(installPrefix, opts));
-                }
             }
         });
     }
@@ -1368,6 +1378,7 @@ export abstract class CMakeDriver implements vscode.Disposable {
      * The list of generators CMake supports as of 3.21
      */
     private readonly cmakeGenerators = [
+        "Visual Studio 18 2026",
         "Visual Studio 17 2022",
         "Visual Studio 16 2019",
         "Visual Studio 15 2017",
@@ -1448,6 +1459,14 @@ export abstract class CMakeDriver implements vscode.Disposable {
         if (!hasExportCompileCommands) {
             const exportCompileCommandsValue = util.cmakeify(exportCompileCommandsFile);
             expandedArgs.push(`-DCMAKE_EXPORT_COMPILE_COMMANDS:${exportCompileCommandsValue.type}=${exportCompileCommandsValue.value}`);
+        }
+        // Only use the installPrefix config if the user didn't
+        // provide one via the preset or configureArgs
+        const hasInstallPrefix = Object.prototype.hasOwnProperty.call(presetCacheVariables, 'CMAKE_INSTALL_PREFIX')
+            || expandedArgs.some(arg => arg.startsWith('-DCMAKE_INSTALL_PREFIX'));
+        if (!hasInstallPrefix && this.installDir) {
+            const installPrefixValue = util.cmakeify(this.installDir);
+            expandedArgs.push(`-DCMAKE_INSTALL_PREFIX:${installPrefixValue.type}=${installPrefixValue.value}`);
         }
         return expandedArgs;
     }
@@ -1933,7 +1952,7 @@ export abstract class CMakeDriver implements vscode.Disposable {
         } else {
             buildPreset.__targets = buildPreset.targets;
         }
-        const args = preset.buildArgs(buildPreset, this.config.buildArgs, this.config.buildToolArgs);
+        const args = preset.buildArgs(buildPreset, this.config.buildArgs, this.config.buildToolArgs, this.config.numJobsForPreset);
         const initialEnvironment = EnvironmentUtils.create(buildPreset.environment);
         const build_env = await this.getCMakeBuildCommandEnvironment(initialEnvironment);
         const expanded_args_promises = args.map(async (value: string) => expand.expandString(value, { ...this.expansionOptions, envOverride: build_env }));
@@ -2017,7 +2036,8 @@ export abstract class CMakeDriver implements vscode.Disposable {
         const buildcmd = await this.getCMakeBuildCommand(targets);
         if (buildcmd) {
             let outputEnc = this.config.outputLogEncoding;
-            if (outputEnc === 'auto') {
+            const isAutoEncoding = outputEnc === 'auto';
+            if (isAutoEncoding) {
                 if (process.platform === 'win32') {
                     outputEnc = await codepages.getWindowsCodepage();
                 } else {
@@ -2034,7 +2054,7 @@ export abstract class CMakeDriver implements vscode.Disposable {
                     }
                 }
             } else {
-                const exeOpt: proc.ExecutionOptions = { environment: buildcmd.build_env, outputEncoding: outputEnc };
+                const exeOpt: proc.ExecutionOptions = { environment: buildcmd.build_env, outputEncoding: outputEnc, useAutoEncoding: isAutoEncoding };
                 this.cmakeBuildRunner.setBuildProcess(this.executeCommand(buildcmd.command, buildcmd.args, consumer, exeOpt));
             }
             const result = await this.cmakeBuildRunner.getResult();